git-svn-id: https://kbengine.svn.sourceforge.net/svnroot/kbengine@207 d26bdfbe-e24b-4f15-971f-8e0596e1bc08

+++ kbe/src/lib/network/network_interface.hpp
@@ -45,7 +45,7 @@ public:
 	typedef std::map<Address, Channel *>	ChannelMap;
 	
 	NetworkInterface(EventDispatcher * pMainDispatcher, NetworkInterfaceType interfaceType,
-		uint16 listeningPort = 0, const char * listeningInterface = "");
+		int32 listeningPort = 0, const char * listeningInterface = "");
 	~NetworkInterface();
 	
 	void attach(EventDispatcher & mainDispatcher);
